fanotify: Add support for fanotify mode
With command line arg --fanotify, the mount point argument is ignored,
FUSE filesystem is not mounted, and cache entires are populated on first
access to the cache data directory.

readdir on the cache directory is required before an entry inside the
cache directory can be accessed, because readdir creates the cache
entry and read populates the cache entry with data.

This mode implies --cache.