fanotify: Crash safe cache write
When we are marking a file as fully cached, the cache should be
fsynced before we set the persistent ignore mask, because after
crash we will not get a permission event to re-fill the cache.

Regardless of fsync, the last cache close must wait for all
background download threads to complete before storing cache
state to disk.