fanotify: Add support for experimental FAN_LOOKUP_PERM
That event can populate all directories in a path when trying
to access a path with unpopulated parents.

To avoid deadlock when handling FAN_LOOKUP_PERM and accessing
the cache data dir, convert all cache data dir access syscalls
to xxxat() sysycalls and use the O_PATH dirfd we got from fanotify
on the first event to perform all access to the cache data dir.