fanotify: Deny user modifications to partially cached files and dirs
Subscribe to all pre modify permission events in mark mount
to deny modification to any of the cached objects that are not
fully populated.

Modification do not wait until file data is downloaded. If cache is
not fully populated - the modification fails.

Modifications to directory content are allowed only not via the watched
mount or by using the fd from the pre modify permission event as the
dirfd argument to syscalls (e.g. mkdirat()), which is how directories
are being populated on access.

Changing attributes is allowed also for unpopulated files and
directories (e.g. chown -R), except for changing size (truncate) and
mtime (touch).