fanotify: Deny user modifications to partially cached files
Subscribe to pre modify permission events in mark mount to deny
modification to any of the cached files that are not fully populated.

Modification do not wait until file data is downloaded. If cache is
not fully populated - the modification fails.

Modifications to file content is allowed only not via the watched mount
or by using the event->fd from the pre modify permission event for
writing content, which is how files are being populated on access.

Changing attributes is allowed also for unpopulated files (e.g. chown),
except for changing size (truncate) and mtime (touch) attributes.