fanotify: Add change tracking group
We have one fanotify group handling cache populate on pre-access events.
Create another group to handle change tracking on pre-change events.

The reason that we need two independent group is because the criteria
for adding and removing an ignore mask is different.

For cache populate the ignore mask is set after populating cache and
removed on cache evict.

For change tracking the ignore mask is set after change is recorded and
all ignore marks will be flushed on checkpoint.

Currently, the change tracking group denies all changes, because change
recording is not implemented.

Soon it will record change for pre-modify and pre-attrib, but it will
still deny create, delete and rename of cached files.